count 자료형 실수로 변경

+++ src/main/java/com/zolPro/yoriLab/domain/IngredientAmount.java
@@ -21,7 +21,7 @@ public class IngredientAmount {
     @JoinColumn
     private Food food;
 
-    private Integer count;
+    private Double count;
 
     private String unit;
 }
